refactor: split config.rs into domain-specific configuration modules (#981)
Split src/config.rs (587 lines) into a src/config/ module directory with
four focused submodules:
- helpers.rs: boolean env var parsing utilities
- user_facing.rs: user-controlled settings (GPU, streaming, watchdog, etc.)
- observability.rs: timing, profiling, GPU metrics
- detection.rs: threshold tuning overrides

All public API remains accessible via re-exports in mod.rs.
No functional changes.
